Store adventure ID, not object, in an attribute
When accessing the Adventure object directly, we have no QuerySet and
so cannot call update(), instead we have to call save(). Storing the
adventure ID means we can access the Adventure via filter(). This
returns a QuerySet, so we can call update().

Calling update() instead of save() is more efficient and avoids a race
condition wherein an object is changed in the database between
retrieval and saving.